Issue #29 DockerFile for MACOS compatibility commands
Problem:
We were trying to use this ECO sonar solution for a client and our client application requires MAC OS for development purpose. After cloning the ECOSONAR api, we were not able to run and deploy the application to docker as the current DOCKER file is having windows specific commands.

Solution:
After working on the file and debugging on this issue, i have done necessary modifications to the current DOCKER File with MAC OS command to incorporate MAC OS execution and deployment.